Moved default ctor of NodeComposer to a factory.
The reason for this is that the default NodeComposer<T> constructor was
really busy, and actually built up a big object graph, which isn't
something I would expect a constructor to do. Particularly when taking
into account the (previously) protected constructor overload that took an
ISpecimenBuilder as argument - that constructor overload doesn't build a
big object graph, but instead accepts the argument as is. The constructors
were inconsistent, which is why I originally made the second one
protected; still, protected is also a sort of public, so there was plenty
room for POLA violation with that design.
Moving the behavior of the former default constructor to a static factory
method seems to be a more honest API because a Create method sounds like
something which is a bit more 'busy'. This is all subjective, so YMMV.