Deprecated CompositeSpecimenBuilder.Builders prop,
(sort of). However, the intention isn't to completely take away this
property, but rather change its return type to
IEnumerable<ISpecimenBuilder>. Since this is a breaking change, it must be
done in v3. When this is done, the property should remain with the new
return type, and the [Obsolete] attribute removed.