Added returned disposable to list of disposables.
The reason is that since the signature of the Compose method gives the
caller no hint that an IDisposable instance is returned, it's the
responsibility of the class itself to trakc and properly dispose of the
returned instance.
Since DisposableTracker alread implements IDisposable, callers should be
aware that they must dispose of the original instance sooner or later, and
when that happens, the returned instance will also be disposed of.