Added EditorBrowsableState.Never to Freeze,
because people always think it does what the Inject method does. Although
I've never myself used this Freeze method as intended, theoretically, it's
a feature, so shouldn't be removed.
However, I don't think anyone actually uses this method correctly, so
hiding it with EditorBrowsableState.Never seems like the best option. If
you really need it, it's still there, but in normal cases you shouldn't
see it in IntelliSense.