Changed BehaviorRoot to a Decorator,
in order to favor composition over inheritance. This makes BehaviorRoot a
pure marker instead of having the mixed responsibility of also acting as a
Composite for other ISpecimenBuilders. When similar changes have also been
applied to other nodes deriving from CompositeSpecimenBuilder, this would
theoretically make it easier to change an entire object graph and, e.g.
replace all Composites with a parallelized version.