Refactored IsTarget to a slightly faster impl.
The problem with IsTarget was that it searched through the graph to find
the target every time it was invoked, and since this predicate was invoked
recursively, this took a long time because it caused loop within loops.

The new implementation is more complicated, but the private
TargetSpecification class holds the target directly, which means that it's
only necessary to find the target once, every time the predicate must be
used in a call to ReplaceNodes or SelectNodes.

The same not particularly rigorous tests as described in the last commit
indicates a 6 % performance improvement.

Before:
========== Total Tests: 2965 passed, 0 failed, 0 skipped, took 10,11 seconds ==========
========== Total Tests: 2965 passed, 0 failed, 0 skipped, took 10,29 seconds ==========
========== Total Tests: 2965 passed, 0 failed, 0 skipped, took 10,08 seconds ==========
========== Total Tests: 2965 passed, 0 failed, 0 skipped, took 10,22 seconds ==========
========== Total Tests: 2965 passed, 0 failed, 0 skipped, took 09,92 seconds ==========

After:
========== Total Tests: 2965 passed, 0 failed, 0 skipped, took 9,34 seconds ==========
========== Total Tests: 2965 passed, 0 failed, 0 skipped, took 9,51 seconds ==========
========== Total Tests: 2965 passed, 0 failed, 0 skipped, took 9,58 seconds ==========
========== Total Tests: 2965 passed, 0 failed, 0 skipped, took 9,56 seconds ==========
========== Total Tests: 2965 passed, 0 failed, 0 skipped, took 9,35 seconds ==========