Changed two calls of Single() to First().
While I think that Single() is more correct, because it will implicitly
check the assertion that there's only a single matching item, it's
theoretically slower because it MUST traverse the entire sequence, whereas
First() stops at the first match it finds.
I can't measure much of a performance difference, but since the TargetMemo
property from the previous commit already already took the plunge and
replaced Single() with First(), I think it makes sense to do this
consistently, as it may provide a performance improvement in scenarios
where users have larger sets of customizations.