made RecursionGuard a concrete class,
instead of an abstract class. The goal is to refactor towards composition
instead of inheritance. Ultimately, this means getting rid of the three
child classes currently deriving from RecursionGuard.

The first step is to make RecursionGuard a concrete class, but that also
means that some abstract members have been changed into virtual.
Currently, they aren't implemented, but that comes next.