Extracted a helper method,
hopefully in order to make the code more readable. You may think it was
more readable before, but I was concerned about the reversal of null
checks: some of the null checks were checking for null, whereas others
were checking for 'not null'. I was concerned that a casual reader might
not notice this (because I actually made that mistake while implementing).