Replace two xml comment instances of 'OmitOnRecursionGuard' with 'RecursionGuard'
The use of OmitOnRecursionGuard class seems to have been replaced with the combined use of RecursionGuard and OmitOnRecursionHandler. The class level xml comments for the OmitOnRecursionBehavior class, as well as its Transform method are updated with this commit to replace the word 'OmitOnRecursionGuard' with 'RecursionGuard' in a total of [2] places.