Perf improvement in RandomNumericSequenceGenerator
The runtime of each individual call to Create on
RandomNumericSequenceGenerator increases with the amount of calls. This is
caused by the check for dublicate numbers that stores already generated
longs in a List<T>. This problem is solved by changing List<T> to
HashSet<T>.

See issue #186

While no tests exists that verifies the performance impact of repeated
calls to Create on the same instance of RandomNumericSequenceGenerator
this is the result that @moodmosaic got when running all the tests before
and after the change.

Before (List<T>):
2940 passed, 0 failed, 0 skipped, took 16.05 seconds (xUnit.net 1.8.0 build 1549).
2940 passed, 0 failed, 0 skipped, took 19.96 seconds (xUnit.net 1.8.0 build 1549).
2940 passed, 0 failed, 0 skipped, took 16.46 seconds (xUnit.net 1.8.0 build 1549).

After (HashSet<T>):
2940 passed, 0 failed, 0 skipped, took 13.20 seconds (xUnit.net 1.8.0 build 1549).
2940 passed, 0 failed, 0 skipped, took 12.92 seconds (xUnit.net 1.8.0 build 1549).
2940 passed, 0 failed, 0 skipped, took 14.00 seconds (xUnit.net 1.8.0 build 1549).