Refactored Parameter-Member matching implementation,
by renaming SemanticReflectionVisitor and SemanticComparisonValue classes to
NameAndTypeCollectingVisitor and NameAndType respectively, in order to better
communicate their intended purpose and behaviour. The default comparison of
NameAndType is now more idiomatic given it's name. Additionally, the
SemanticElementComparer was renamed to ReflectionVisitorElementComparer, and
the IEqualityComparer<T> implementation can be passed into the constructor. This
allows for easier future change (which is expected). This commit now means all
existing tests are passing again.