Removed a test that no longer always passes.
This test must be a remnant of AutoFixture 2.x, because in that version
range, it was supposed to always pass. In AutoFixture 3.x, OTOH, it may
occasionally fail, which I've seen happen once. However, this only happens
rarely.
The test is incorrect for AutoFixture 3, so therefor I remove it. It's not
a breaking change, because the behaviour of AutoFixture doesn't change by
removing the test.