Configured NUnit2 projects to sign the assemblies.
Apparently, this was missed when the projects were originally created. I
must admit that I didn't even know that this was necessary, but I think
it's because we need to exclude some projects from being signed. See
commit 91a2423c967440a761e8728132f3728907206bd4 for more details.

This particular commit is in response to issue #423. Only after build and
publication of NuGet packages will I know for sure if it works. I hope
mentioning the issue here doesn't close it.