Refactored reflection code to generic class,
in order to avoid some of the readability (and performance?) overhead of
using exclusively reflection in order to perform the desired work.
Instead, the reflection hit is paid only once when the private
Filler<TKey, TValue> object is created. Hopefully, the code in the private
Fill method is easier to follow.