Moved InvariantCultureGenerator out of engine,
because that part of the engine (DefaultPrimitiveBuilders) sits inside of
AutoPropertiesTarget, which means that Fixture attempts to assign values
to writeable properties after creation. It turns out that CultureInfo is
an extremely mutable type, so AutoFixture chokes on some of the writeable
properties.

This also illustrates that passing unit tests aren't always enough; a
single Facade Test could have caught this problem when
InvariantCultureGenerator was originally added.