Made an InlineAutoDataAttribute constructor overload protected.
InlineAutoDataAttribute has a constructor overload that takes an
AutoDataAttribute as an argument.

As #457 implicitly points out, there's no way to use this constructor
overload when using the InlineAutoDataAttribute to annotate tests.

This constructor overload only exists to enable a derived class to supply
a custom, derived instance of AutoDataAttribute, thereby customizing the
behaviour of the derived attribute.

Since this constructor overload only exists to enable inheritance, we are
now changing its accessibility level to protected.