Supported implicit conversions in With expressions
It turns out that the support for conversions in `With` expressions added
previously caused a regression that we've never tested for: when an
implicit conversion exists (implemented by op_Implicit), the `With`
expression used to work, but after explicitly attempting to convert using
Convert.ChangeType, this doesn't always work (e.g. for Nullable<T>).

There seems to be no easy way to use Reflection to figure out if an
implicit conversion already exists, so the pragmatic solution seems to be
to attemp making the conversion in a try block, and then retry with the
conversion if the first attempt fails.

If someone learns of a better way to test for this without resorting to
exception handling, a pull request will be welcome.