Deleted orphaned file.
This file was removed in 53e22adc35a232be8e59841de9b83b8c02275136, but it
seems I subsequently accidentally reintroduced it to the repository in
2c81087a99d512f69ef239a03dde4200d4c99c65, when moving files in to the /Src
directory. This was in the spring of 2010, and I think we were using
Mercurial at the time. Apparently I didn't fully understand how to use Hg
at the time.