Deleted orphaned file.
Apparently, in d9bf1053d04742d588fcf59c76035bb4c6241325 I meant to move
Seed.cs to the Kernel directory, but instead of moving the file, I copied
it. This was back in early 2010, and I think we were using Mercurial at
the time. It looks like I didn't quite understand how to use it.