Renamed `TypeExtensions` to `TypeEnvy`.
The class contains methods that are not extension methods, so it makes
poor sense to name it with the Extensions suffix. The Envy suffix takes
its name from the Feature Envy code smell. Essentially, all extension
methods are code smells, but it can hardly be helped, since it's
impossible to add methods to existing BCL types otherwise.