Made an AutoDataAttribute constructor overload protected.
AutoDataAttribute has a constructor overload that takes an IFixture as an
argument.

As #532 points out, there's no way to use this constructor overload when
using the AutoDataAttribute to annotate tests.

This constructor overload only exists to enable a derived class to supply
a custom, derived instance of AutoDataAttribute, thereby customizing the
behaviour of the derived attribute.

Since this constructor overload only exists to enable inheritance, we are
now changing its accessibility level to protected.