SYSTEM_NET_MAIL and SYSTEM_RUNTIME_SERIALIZATION
System.Net.Mail APIs, used by AutoFixture in the MailAddressGenerator, are currently only available on the full .net framework (e.g. net45), so we wrap them entirely in an `#if SYSTEM_NET_MAIL` directive.

Similarly, serialization (including SerializableAttribute and binary remoting serialization) is only available on the full .net framework, so we use `#if SYSTEM_RUNTIME_SERIALIZATION` to allow us to compile on targets that don't have those features.