Rework the way of suppressing a particular member auto population
Previously if it was needed to suppress AutoProperties command from
populating a particular member we inserted a global Omitter.
That solution was bad because Omitter was registered outside
of the Customization scope and affected unrelated requests.

Current solution is to configure the `Specification` field of the
AutoPropertiesCommand<> and exclude the suppressed fields/props.