Refactor TerminatingWithPathSpecimenBuilder to avoid Tracer usage
Change implementation to avoid usage of the TracingBuilder.
That allows to make code much simpler and clearer. Now all the
requests are being recorded and it's not possible to filter them.
That should not be an issue as this information is used and
displayed in in exceptional situations only.

Adjusted tests as design was changed heavily.