Disable parallel test execution for RhinoMock integration tests
It appears that RhinoMock is not thread safe and our xUnit tests
could fail due to parallel test execution. The issue is intermittent
and outcome might be different - from exception to complete hang.
You can read more information about this issue here:
http://blog.smithfamily.dk/post/Thread-safe-version-of-Rhino-Mocks

This is old version of RhinoMock that is no longer supported, so
the only option we have is to disable the parallel test execution.