Improve Fixture.OmitAutoProperties internal approach
The previous approach was based on removal and insertion of the
postprocessor node which fills up the properties. The new approach
is to always keep the postprocessor, but instead enable/disable
it using the different specifications.

This approach makes the code much simpler and provides better
current value lookup performance.