Create RangedRequest in RangeAttributeRelay
The previous approach when we created RangedNumberRequest directly
caused issues as Range attribute might be applied to a non-number
types (e.g. Enums). In that case we failed with exception.

The new architecture adds a relay from RangedRequest to a
RangedNumberRequest if type is a number. Otherwise, it gives a
chance to other builders to handle other request types.