Rework random generator to use LCG
The previous implementation sufferred from a few major issues.
The first issue is that it failed with OOM if large number
of numeric requests are issued. Another issue is that
generation time of the latest elements in a range
could take large amount of time, as it works in a loop.

The new implementation uses the Linear Congruential Generator
approach where the properties of prime numbers are used to
generate non-repeating sequence on a particular range. While
we are still generating numbers in a loop ignoring out-of-range
values, the predictability of this approach is much better
because of the values distribution. If distribution were uniform,
we'd ignore at most half of the generated numbers. In reality
the distibution is worse, however still good enough to rely on it.