Skip "copy" constructors in all the queries (#937)
Changed the EnumerableFavoring-, ArrayFavoring-, ListFavoring-,
Greedy- and ModestConstructorQuery behavior to skip the constructors
which contain parameters of the enclosing type.

AutoFixture is unable to activate instances using such constructors
as it leads to the recursion and subsequent failure. Therefore, to
make the build-in heuristic more robust, we simply omit those ones.