Fix spelling and grammar and tweak wording
I noticed the "instanes" on line 46, then the other misspellings. Then "like that is" hit my ear funny, so I took the controversial step of replacing with "as was". Then I though someone might think "demanded" was harsh, so I took the super controversial step of softening it.

As usual, happy to make any amendments.