Deprecate the MapCreateManyToEnumerable customization
This customization leads to the recursion if client forgets to
add a customization for the particular IEnumerable<T>. This kind of
behavior is really confusing to bundle it with the library.
Instead, clients are advised to use the code snippet to customize
the MultipleRequest requests for the desired sequence types only.