Suggestions from codereview: Added a test to FixtureTests, to verify that #1183 is fixed as suggested here: https://github.com/AutoFixture/AutoFixture/pull/1187#discussion_r447634951
This required making sure that the outcome space of the regex pattern is great enough. Xeger seems to favor shorter strings, and this gave collisions when generating RegularExpressionValidatedType . So minimum length of the pattern in this type was lengthened from 1 to 20.

Removed assertion roulette in RegularExpressionGeneratorTests.CreateMultipleWithRegularExpressionRequestReturnsDifferentResults(...)
according to suggestion: https://github.com/AutoFixture/AutoFixture/pull/1187#discussion_r447633342